Moved App namespace up a level to ClientApp
