+++ src/Hypzeh.Web/ClientApp/components/layout/Panel/index.jsx
