+++ src/Hypzeh.Web/ClientApp/components/projects/Smallify/components/pages/Home/index.jsx
