+++ src/Hypzeh.Web/ClientApp/components/projects/Smallify/index.jsx
