+++ src/Hypzeh.Web/ClientApp/components/shared/Icon/index.jsx
