+++ src/Hypzeh.Web/ClientApp/contexts/view/view-actions.js
