+++ src/Hypzeh.Web/ClientApp/contexts/view/view-reducer.js
