+++ src/Hypzeh.Web/ClientApp/contexts/view/view-state.js
