+++ src/Hypzeh.Web/ClientApp/contexts/view/view-types.js
