+++ src/Hypzeh.Web/ClientApp/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/App';
+import App from './App';
 
 const root = document.createElement('div');
 root.setAttribute('id', 'root');
