+++ src/Hypzeh.Web/ClientApp/utils/navigation.js
