+++ src/Hypzeh.Web/ClientApp/utils/style/global-style.js
